Move Auth0 into package folder, migrate to vapor 4 and update Package.swift
