+++ Package.swift
@@ -8,6 +8,8 @@ let package = Package(
     ],
     products: [
         .library(name: "ImperialCore", targets: ["ImperialCore"]),
+        .library(name: "ImperialAuth0", targets: ["ImperialCore", "ImperialAuth0"]),
+        .library(name: "ImperialDiscord", targets: ["ImperialCore", "ImperialDiscord"]),
         .library(name: "ImperialDropbox", targets: ["ImperialCore", "ImperialDropbox"]),
         .library(name: "ImperialFacebook", targets: ["ImperialCore", "ImperialFacebook"]),
         .library(name: "ImperialGitHub", targets: ["ImperialCore", "ImperialGitHub"]),
@@ -16,9 +18,10 @@ let package = Package(
         .library(name: "ImperialKeycloak", targets: ["ImperialCore", "ImperialKeycloak"]),
         .library(name: "ImperialMicrosoft", targets: ["ImperialCore", "ImperialMicrosoft"]),
         .library(name: "ImperialShopify", targets: ["ImperialCore", "ImperialShopify"]),
-        .library(name: "ImperialDiscord", targets: ["ImperialCore", "ImperialDiscord"]),
         .library(name: "Imperial", targets: [
             "ImperialCore",
+            "ImperialAuth0",
+            "ImperialDiscord",
             "ImperialDropbox",
             "ImperialFacebook",
             "ImperialGitHub",
@@ -26,8 +29,7 @@ let package = Package(
             "ImperialGoogle",
             "ImperialKeycloak",
             "ImperialMicrosoft",
-            "ImperialShopify",
-            "ImperialDiscord"
+            "ImperialShopify"
         ]),
     ],
     dependencies: [
@@ -42,6 +44,8 @@ let package = Package(
                 .product(name: "JWTKit", package: "jwt-kit"),
             ]
         ),
+        .target(name: "ImperialAuth0", dependencies: ["ImperialCore"]),
+        .target(name: "ImperialDiscord", dependencies: ["ImperialCore"]),
         .target(name: "ImperialDropbox", dependencies: ["ImperialCore"]),
         .target(name: "ImperialFacebook", dependencies: ["ImperialCore"]),
         .target(name: "ImperialGitHub", dependencies: ["ImperialCore"]),
@@ -50,7 +54,6 @@ let package = Package(
         .target(name: "ImperialKeycloak", dependencies: ["ImperialCore"]),
         .target(name: "ImperialMicrosoft", dependencies: ["ImperialCore"]),
         .target(name: "ImperialShopify", dependencies: ["ImperialCore"]),
-        .target(name: "ImperialDiscord", dependencies: ["ImperialCore"]),
         .testTarget(name: "ImperialTests", dependencies: ["ImperialCore", "ImperialShopify"]),
     ]
 )
