+++ Sources/ImperialAuth0/Auth0.swift
@@ -1,3 +1,4 @@
+@_exported import ImperialCore
 import Vapor
 
 public class Auth0: FederatedService {
@@ -6,18 +7,18 @@ public class Auth0: FederatedService {
     
     @discardableResult
     public required init(
-        router: Router,
+        routes: RoutesBuilder,
         authenticate: String,
-        authenticateCallback: ((Request)throws -> (Future<Void>))?,
+        authenticateCallback: ((Request) throws -> (EventLoopFuture<Void>))?,
         callback: String,
         scope: [String] = [],
-        completion: @escaping (Request, String)throws -> (Future<ResponseEncodable>)
-    )throws {
+        completion: @escaping (Request, String) throws -> (EventLoopFuture<ResponseEncodable>)
+    ) throws {
         self.router = try Auth0Router(callback: callback, completion: completion)
         self.tokens = self.router.tokens
         
         self.router.scope = scope
-        try self.router.configureRoutes(withAuthURL: authenticate, authenticateCallback: authenticateCallback, on: router)
+        try self.router.configureRoutes(withAuthURL: authenticate, authenticateCallback: authenticateCallback, on: routes)
         
         OAuthService.register(.auth0)
     }
